+++ R/data_test.R
@@ -18,7 +18,7 @@
 #'
 #' @examples
 #' library(lavaan)
-#' data(data_med)
+#' data(data_test_3_factor_3_item)
 #' mod <-
 #' "
 #' fx =~ x1 + x2 + x3