+++ R/data_test4_4obvs.R
@@ -1,4 +1,4 @@
-#' @title Test Dataset: 4obvs
+#' @title Test Dataset: 4 Observed Variables
 #'
 #' @description A dataset for testing.
 #'
@@ -13,7 +13,7 @@
 #'
 #' @examples
 #' library(lavaan)
-#' data(data_med_4vars)
+#' data(data_test_4obvs)
 #' mod <-
 #' "
 #' fm1 ~~ fm2