+++ R/plot_ptables.R
@@ -1,14 +1,16 @@
-#' @title Plot Functions for a List of Parameter Tables
+#' @title Plot Functions for a List of Models
 #'
 #' @description Various plot functions
-#' for the output of [eq_models()] and
+#' for the output of [eq_models()],
+#' [eq_df_models()], and
 #' similar functions.
 #'
 #' @details
 #' The functions provide different ways
 #' to visualize the models. Basic knowledge
 #' of [semPlot::semPaths()] from the
-#' package `semPlot` is required.
+#' package `semPlot` is required to
+#' customize the plots.
 #'
 #' @seealso See [eq_models()] on the
 #' type of output supported.
@@ -59,8 +61,8 @@
 #'   eq_out_1,
 #'   original_model = fit1,
 #'   layout = layout_i,
-#'   label.cex = 1.8,
-#'   sizeMan = 11,
+#'   label.cex = 1.5,
+#'   sizeLat = 15,
 #'   edge.width = 5,
 #'   asize = 5,
 #'   structural = TRUE,
@@ -82,8 +84,18 @@ NULL
 #' from [semPlot::semPaths()], of the
 #' class `partables_plots`.
 #'
+#' @param object If it is a list of models
+#' (parameter tables), such as the output
+#' of [eq_models()], it will be used
+#' as the value for `partables`. If
+#' it is a `partables_plots` object
+#' (i.e., an output of [partables_plots()],
+#' then it will be updated with any new
+#' values for other arguments.
+#'
 #' @param partables It should be a list of
-#' parameter tables, such as the output
+#' models in the form of parameter tables,
+#' such as the output
 #' of [eq_models()] or [eq_df_models()].
 #'
 #' @param original_model The model to
@@ -91,7 +103,8 @@ NULL
 #' compared to. If `NULL`, then the
 #' plots will be generated without checking
 #' for differences between a model and
-#' `original_model`.
+#' `original_model`. Model comparison is
+#' conducted by [model_diff()].
 #'
 #' @param ... Optional arguments to be
 #' passed to [semPlot::semPaths()]. They
@@ -103,7 +116,7 @@ NULL
 #' to a model in `partables`. For now,
 #' two settings are supported: `color`
 #' for the color and `width` for the
-#' width of an edge (arrow).
+#' width of an edge (arrow/path).
 #'
 #' @param fix_pars_fixed_zero If `TRUE`,
 #' parameters fixed to zero will be
@@ -114,9 +127,9 @@ NULL
 #' fixed to zero. For now,
 #' two settings are supported: `color`
 #' for the color and `width` for the
-#' width of an edge (arrow). Setting the
-#' width to zer0, the default, effectively
-#' hide an arrow.
+#' width of an edge (arrow/path). Setting the
+#' width to zero, the default, effectively
+#' hiding an arrow/path.
 #'
 #' @param exclude_original_model If `TRUE`
 #' and `original_model` is set, the output
@@ -134,8 +147,9 @@ NULL
 #' @rdname plot_partables
 #' @export
 partables_plots <- function(
-  partables,
+  object,
   ...,
+  partables = NULL,
   original_model = NULL,
   auto_node_color = TRUE,
   par_diff_settings = list(
@@ -147,8 +161,74 @@ partables_plots <- function(
               color = "white",
               width = 0
             ),
-  exclude_original_model = TRUE
+  exclude_original_model = FALSE
 ) {
+
+  is_update <- FALSE
+
+  if (inherits(object, "partables_plots")) {
+
+    # ==== Update a previous output ====
+
+    is_update <- TRUE
+    args_names <- names(formals())
+
+    # Retrieve stored argument values
+
+    args0 <- attr(object, "args")
+    call_new <- as.list(match.call())[-1]
+    # object is just a placeholder,
+    # it can be previous output or
+    # a partables for a new set of plots
+    call_new$object <- NULL
+    ddd_old <- args0[!(names(args0) %in% args_names)]
+    eee_old <- args0[names(args0) %in% args_names]
+
+    # Get the new values and evaluate them
+    # TODO:
+    # - How to handle arguments not named?
+    ddd_new <- call_new[!(names(call_new) %in% args_names)]
+    eee_new <- call_new[names(call_new) %in% args_names]
+    ddd_new <- lapply(ddd_new, eval, envir = parent.frame())
+    eee_new <- lapply(eee_new, eval, envir = parent.frame())
+
+    ddd <- utils::modifyList(
+      ddd_old,
+      ddd_new,
+      keep.null = TRUE
+    )
+    eee <- utils::modifyList(
+      eee_old,
+      eee_new,
+      keep.null = TRUE
+    )
+
+    # modifyList does not handle an argument of list as desired
+    eee$partables <- eee_new$partables %||% eee_old$partables
+    eee$original_model <- eee_new$original_model
+
+    # Set the argument values
+    # TODO:
+    # - Is there a better way to do this?
+    partables <- eee$partables
+    original_model <- eee$original_model
+    auto_node_color <- eee$auto_node_color
+    par_diff_settings <- eee$par_diff_settings
+    fix_pars_fixed_zero <- eee$fix_pars_fixed_zero
+    par_fixed_zero_settings <- eee$par_fixed_zero_settings
+    exclude_original_model <- eee$exclude_original_model
+    # eee is not needed after this line
+
+  } else if (is_partables(object)) {
+
+    # ==== A new set of plots ====
+
+    partables <- object
+    object <- NULL
+    ddd <- list(...)
+
+  }
+
   if (!is_partables(partables)) {
     stop("partables is not of a supported type.")
   }
@@ -195,7 +275,7 @@ partables_plots <- function(
     pt = partables,
     pars_diff = pt_pars_diff,
     MoreArgs = c(
-      list(...),
+      ddd,
       list(
         auto_node_color = auto_node_color,
         fix_pars_fixed_zero = fix_pars_fixed_zero,
@@ -207,37 +287,113 @@ partables_plots <- function(
     USE.NAMES = TRUE
   )
 
+  out_org <- NULL
+
   if (exclude_original_model &&
       !is.null(original_model)) {
+
+    # ==== Exclude the original model from the output? ====
+
     tmp <- partables %pt_in% as_eq_partables(original_model)
     if (any(tmp)) {
       out_org <- out[tmp]
       out <- out[!tmp]
-      attr(out, "original_model") <- out_org
+      # attr(out, "original_model") <- out_org
     }
   }
 
+  if (!is.null(original_model)) {
+
+    # ==== Store the plot of the original model ====
+
+    if (is.null(out_org)) {
+      out_org <- do.call(
+        partables_plots_internal,
+        c(list(
+            pt = original_model,
+            auto_node_color = auto_node_color,
+            fix_pars_fixed_zero = fix_pars_fixed_zero,
+            par_fixed_zero_settings = par_fixed_zero_settings,
+            par_diff_settings = par_diff_settings
+          ),
+          ddd)
+      )
+    }
+    attr(out, "original_model") <- list(original_model = out_org)
+
+  }
+
+  # ==== Store Args ====
+
+  if (is_update) {
+
+    # ==== Store the updated argument values ====
+
+    args <- c(ddd, eee)
+    attr(out, "args") <- args
+
+  } else {
+
+    # ==== A new call ====
+
+    call_new <- as.list(match.call())
+    call_new[[1]] <- NULL
+    # object is just a placeholder
+    call_new$object <- NULL
+    call_new <- lapply(
+      call_new,
+      eval,
+      envir = parent.frame()
+    )
+
+    # Update the default values
+    call0 <- formals()
+    # object is just a placeholder
+    call0$object <- NULL
+    call0$... <- NULL
+    call0 <- utils::modifyList(
+      call0,
+      call_new,
+      keep.null = TRUE
+    )
+    # This is necessary because modifyList does not
+    # handle argument values of names list as desired
+    call0$partables <- partables
+    call0$original_model <- original_model
+    attr(out, "args") <- call0
+
+  }
+
+  # ==== Set class ====
+
   class(out) <- c("partables_plots", class(out))
   out
+
 }
 
-#' @return The `plot` method of the output
-#' of [partables_plots()] return `x`
+#' @return
+#' The `plot` method of the output
+#' of [partables_plots()] returns `x`
 #' invisibly. Called for its side effect.
 #'
 #' @param x The output of [partables_plots()],
 #' a `partables_plots` object.
 #'
-#' @param ... Optional arguments to be
+#' @param ... For [plot.partables_plots()],
+#' these are optional arguments to be
 #' passed to [semPlot::semPaths()].
+#' For [print.partables_plots()], they
+#' are not used.
 #'
 #' @param title_mode What will be used
 #' as the title. If `"digest"`, then
-#' the digest value of a model will be
+#' the digest value of a model, generated
+#' by [digest_partable()], will be
 #' used as its title. If `"name"`, then
-#' its name in `x` will be used. If
+#' its name in `x` will be used, which may
+#' also be the digest value. If
 #' `"none"`, then no title will be drawn
-#' with  the plot.
+#' with the plot.
 #'
 #' @param title_adj Adjust the position
 #' of the title. Increase this value
@@ -283,7 +439,7 @@ partables_plots <- function(
 #' drawn. If `"include"`, the original
 #' model will be drawn as the first model,
 #' along with other models. If `"side_by_side"`,
-#' then number of columns is always two
+#' then the number of columns is always two
 #' and the number of rows is always one.
 #' In each plot, the original model will
 #' be drawn on the left and the other
@@ -307,11 +463,13 @@ plot.partables_plots <- function(
     "Edges:width",
     "Edges:asize"
   ),
-  original_model_mode = c("exclude", "include", "side_by_side")
+  original_model_mode = c("include", "exclude", "side_by_side")
 ) {
   original_model_mode <- match.arg(original_model_mode)
   scale_plots <- match.arg(scale_plots)
   title_mode <- match.arg(title_mode)
+
+  x_org <- x
 
   if (original_model_mode == "side_by_side") {
     nrow <- 1
@@ -330,11 +488,48 @@ plot.partables_plots <- function(
   }
 
   if (has_original_model) {
-    x <- c(original_model_list,
-           x)
+
+    x_digest <- sapply(x, \(x) attr(x, "digest"))
+
+    if (original_model_mode == "exclude") {
+
+      # ==== Exclude the original model, if present ====
+
+      x <- x[!(x_digest %in% original_digest)]
+
+    }
+    if (original_model_mode == "include") {
+
+      # ==== Include the original model ====
+
+      # x may not have the original model
+      # Always insert or move the original model as the first plot
+      x <- x[!(x_digest %in% original_digest)]
+      x <- c(original_model_list,
+            x)
+
+    }
+    if (original_model_mode == "side_by_side") {
+
+      # ==== Side-By-Side ====
+
+      # Insert or move the original model as the first plot
+      x <- x[!(x_digest %in% original_digest)]
+      x <- c(original_model_list,
+            x)
+
+      k0 <- seq_along(x)
+      k1 <- sapply(
+              k0[-1],
+              \(x) c(1, x)
+            )
+      x <- x[k1]
+
+    }
   }
 
   # ==== auto_scale ====
+
   if (scale_plots == "auto") {
     if (((ncol != 1) || (nrow != 1))) {
       scale_i <- scale %||% c(x = ncol / 2, y = nrow / 2)
@@ -359,23 +554,6 @@ plot.partables_plots <- function(
       scale_y = scale_i["y"],
       elements = elements_to_scale
     )
-  }
-
-  if (has_original_model) {
-    if (original_model_mode == "exclude") {
-      x <- x[-1]
-    }
-    if (original_model_mode == "include") {
-      # A placeholder
-    }
-    if (original_model_mode == "side_by_side") {
-      k0 <- seq_along(x)
-      k1 <- sapply(
-              k0[-1],
-              \(x) c(1, x)
-            )
-      x <- x[k1]
-    }
   }
 
   # ==== Plot the model ====
@@ -409,6 +587,34 @@ plot.partables_plots <- function(
     }
   }
   par(parold)
+  invisible(x)
+}
+
+#' @return
+#' The `print` method of the output
+#' of [partables_plots()] returns `x`
+#' invisibly. Called for its side effect.
+#'
+#' @rdname plot_partables
+#' @export
+print.partables_plots <- function(
+  x,
+  ...
+) {
+  x1 <- x
+  class_old <- class(x)
+  i <- which(class_old == "partables_plots")
+  if (length(i) > 0)
+  class(x1) <- class_old[-seq(1, i)]
+  k <- length(x1)
+  x_names <- names(x1)
+  cat("The plot(s) of", k, "model(s):\n")
+  if (!is.null(x_names)) {
+    print(x_names)
+  } else {
+    x_names <- sapply(x, \(x) attr(x, "digest"))
+    print(unname(x_names))
+  }
   invisible(x)
 }
 
@@ -569,6 +775,7 @@ partables_plots_internal <- function(
 
   attr(p_fit, "digest") <- get_digest(pt)
   attr(p_fit, "gen_models_name") <- attr(pt, "gen_models_name")
+  attr(p_fit, "partable") <- pt
 
   p_fit
 }
@@ -599,6 +806,64 @@ scale_plot <- function(
     )
     out$graphAttributes[[xx1[1]]][[xx1[2]]] <-
       out$graphAttributes[[xx1[1]]][[xx1[2]]] * scale1
+  }
+  out
+}
+
+#' @return
+#' The `%p>%` operator returns a
+#' `partables_plots` object, processed
+#' by the right-hand side function.
+#'
+#' @param lhs A `partables_plots` object.
+#'
+#' @param rhs A function call to be applied
+#' to all stored plots. Each plot will
+#' be inserted as the first argument.
+#'
+#' @rdname plot_partables
+#' @export
+`%p>%` <- function(
+  lhs,
+  rhs
+) {
+
+  # A pipe operator to process all stored plots
+
+  rhs_call <- as.list(substitute(rhs))
+  fct <- eval(rhs_call[[1]])
+  out <- lapply(
+    lhs,
+    \(x) {do.call(
+      fct,
+      c(list(x),
+        rhs_call[-1])
+    )}
+  )
+
+  # ==== Has an original model? ====
+
+  original_model <- attr(lhs, "original_model")
+  has_original_model <- !is.null(original_model)
+  if (has_original_model) {
+    out0 <-
+        do.call(
+          fct,
+          c(list(original_model[[1]]),
+            rhs_call[-1])
+      )
+    original_model_modified <- original_model
+    original_model_modified[[1]] <- out0
+  }
+
+  # ==== Finalize the output ====
+
+  out <- overwrite_attributes(
+          out,
+          lhs
+        )
+  if (has_original_model) {
+    attr(out, "original_model") <- original_model_modified
   }
   out
 }