+++ tests/testthat/test-plot_partables.R
@@ -22,6 +22,12 @@ layout_i <- layout_matrix(
   fy = c(2, 3)
 )
 
+layout_j <- layout_matrix(
+  fx = c(1, 1),
+  fm = c(2, 2),
+  fy = c(1, 3)
+)
+
 p1 <- partables_plots(
   pt1,
   layout = layout_i
@@ -39,9 +45,88 @@ p3 <- partables_plots(
   original_model = fit1
 )
 
+p1b <- partables_plots(
+  p1,
+  sizeMan = 20,
+  partables = pt_list_3_obv[4:6],
+  par_diff_settings = list(
+              color = "red",
+              width = 4
+            )
+)
+
+p1c <- partables_plots(
+  p1,
+  sizeMan = 20,
+  original_model = pt_list_3_obv[[2]],
+  partables = pt_list_3_obv[4:6],
+  par_diff_settings = list(
+              color = "red",
+              width = 4
+            ),
+  exclude_original_model = FALSE
+)
+
+
+p2b <- partables_plots(
+  p2,
+  original_model = pt1[[2]],
+  sizeMan = 20,
+  par_diff_settings = list(
+              color = "green",
+              width = 4
+            )
+)
+
+p3b <- partables_plots(
+  p3,
+  layout = layout_j,
+  original_model = pt1[[2]]
+)
+
+expect_no_error(print(p1))
+expect_no_error(print(p2))
+expect_no_error(print(p3))
+
+
 dev.off()
 
+# Check selection
+
+expect_length(have_pars_all(p1, "fx ~ fm"), 1)
+expect_length(have_pars_any(p1, c("fx ~ fm", "fy ~ fx")), 2)
+expect_length(have_pars_none(p1, c("fm ~~ fy")), 1)
+expect_length(must_not_be_y(p1, var = "fx"), 2)
+expect_length(must_be_y(p1, var = "fx"), 1)
+expect_length(must_not_have_paths(p1, y_on_x = "fy ~ fm"), 2)
+expect_length(must_have_paths(p1, y_on_x = "fy ~ fm"), 1)
+
+
 skip("Test in an interactive session.")
+
+# No original model
+plot(
+  p1,
+  ncol = 2,
+  nrow = 2,
+  title_adj = 1
+)
+
+# No original model
+plot(
+  p1b,
+  ncol = 2,
+  nrow = 2,
+  title_adj = 2
+)
+
+# Original model added
+plot(
+  p1c,
+  ncol = 2,
+  nrow = 2,
+  title_adj = 2
+)
 
 plot(
   p2,
@@ -51,11 +136,10 @@ plot(
 )
 
 plot(
-  p2,
+  p2b,
   ncol = 2,
   nrow = 2,
-  title_adj = 1,
-  original_model_mode = "include"
+  title_adj = 2
 )
 
 plot(
@@ -63,7 +147,15 @@ plot(
   ncol = 2,
   nrow = 2,
   title_adj = 1,
-  original_model_mode = "include",
+  original_model_mode = "exclude"
+)
+
+plot(
+  p2,
+  ncol = 2,
+  nrow = 2,
+  title_adj = 1,
+  original_model_mode = "exclude",
   title_mode = "name"
 )
 
@@ -71,7 +163,7 @@ plot(
   p2,
   ncol = 2,
   nrow = 2,
-  title_adj = 1,
+  title_adj = .7,
   original_model_mode = "side_by_side",
   title_mode = "name"
 )
@@ -140,6 +232,14 @@ plot(
   title_args = list(line = 2)
 )
 
-
+plot(
+  p3b,
+  ncol = 3,
+  nrow = 3,
+  scale = 3,
+  original_model_mode = "exclude",
+  title_mode = "name",
+  title_args = list(line = 2)
+)
 
 })